requirements.txt
hyperliquid-python-sdk>=0.4.0
requests>=2.31.0

